@@ -1,80 +1,73 @@
 {
-    "name": "vue-scscms",
-    "version": "1.0.0",
-    "description": "Vue.js and Koa project",
-    "author": "scscms <guangda1234@126.com>",
-    "private": true,
-    "scripts": {
-        "dev": "node build/dev-server.js",
-        "build": "node build/build.js",
-        "server": "nodemon --delay 2000ms -w app.js -w server server-entry.js",
-        "start": "concurrently -k \"npm run server\" \"npm run dev\""
+  "name": "vue-scscms",
+  "version": "2.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "vue-cli-service serve",
+    "server": "nodemon --delay 2000ms -w app.js -w server server-entry.js",
+    "start": "concurrently -k \"npm run server\" \"npm run dev\"",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
+  },
+  "dependencies": {
+    "axios": "^0.18.0",
+    "babel-polyfill": "^6.23.0",
+    "bcryptjs": "^2.4.0",
+    "element-ui": "^2.3.9",
+    "koa": "^2.5.0",
+    "koa-bodyparser": "^4.2.1",
+    "koa-json": "^2.0.2",
+    "koa-jwt": "^3.2.1",
+    "koa-multer": "^1.0.2",
+    "koa-router": "^7.1.1",
+    "koa-static": "^4.0.3",
+    "koa2-history-api-fallback": "^0.0.5",
+    "mysql2": "^1.3.6",
+    "nodemailer": "^4.0.1",
+    "nprogress": "^0.2.0",
+    "vue": "^2.5.17",
+    "vue-quill-editor": "^3.0.6",
+    "vue-router": "^3.0.1",
+    "vuex": "^3.0.1"
+  },
+  "devDependencies": {
+    "@vue/cli-plugin-babel": "^3.0.1",
+    "@vue/cli-plugin-eslint": "^3.0.1",
+    "@vue/cli-service": "^3.0.1",
+    "@vue/eslint-config-standard": "^3.0.4",
+    "babel-core": "^6.26.3",
+    "babel-preset-env": "^1.7.0",
+    "concurrently": "^4.0.1",
+    "http-proxy-middleware": "^0.19.0",
+    "less": "^3.0.4",
+    "less-loader": "^4.1.0",
+    "nodemon": "^1.18.4",
+    "vue-template-compiler": "^2.5.17"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
     },
-    "dependencies": {
-        "axios": "^0.15.3",
-        "babel-polyfill": "^6.23.0",
-        "bcryptjs": "^2.4.0",
-        "element-ui": "^1.4.3",
-        "koa": "^2.5.0",
-        "koa-bodyparser": "^3.2.0",
-        "koa-json": "^2.0.2",
-        "koa-jwt": "^3.2.1",
-        "koa-logger": "^2.0.1",
-        "koa-multer": "^1.0.2",
-        "koa-router": "^7.1.1",
-        "koa-static": "^3.0.0",
-        "koa2-history-api-fallback": "^0.0.2",
-        "log4js": "^2.3.0",
-        "mysql2": "^1.3.6",
-        "nodemailer": "^4.0.1",
-        "nprogress": "^0.2.0",
-        "vue": "^2.3.4",
-        "vue-html5-editor": "^1.1.1",
-        "vue-router": "^2.3.1",
-        "vuex": "^2.3.1"
+    "extends": [
+      "plugin:vue/essential",
+      "@vue/standard"
+    ],
+    "rules": {
+      "no-console": 0
     },
-    "devDependencies": {
-        "autoprefixer": "^6.7.2",
-        "babel-core": "^6.22.1",
-        "babel-loader": "^6.2.10",
-        "babel-preset-env": "^1.6.1",
-        "babel-preset-es2015": "^6.24.1",
-        "babel-register": "^6.26.0",
-        "caniuse-lite": "^1.0.30000701",
-        "chalk": "^1.1.3",
-        "concurrently": "^3.5.0",
-        "connect-history-api-fallback": "^1.3.0",
-        "copy-webpack-plugin": "^4.0.1",
-        "css-loader": "^0.28.4",
-        "eventsource-polyfill": "^0.9.6",
-        "express": "^4.14.1",
-        "extract-text-webpack-plugin": "^2.0.0",
-        "file-loader": "^0.11.1",
-        "friendly-errors-webpack-plugin": "^1.1.3",
-        "has-flag": "^2.0.0",
-        "html-webpack-plugin": "^2.28.0",
-        "http-proxy-middleware": "^0.17.3",
-        "less": "^2.7.2",
-        "less-loader": "^4.0.5",
-        "nodemon": "^1.11.0",
-        "opn": "^4.0.2",
-        "optimize-css-assets-webpack-plugin": "^1.3.0",
-        "ora": "^1.2.0",
-        "rimraf": "^2.6.0",
-        "semver": "^5.3.0",
-        "shelljs": "^0.7.6",
-        "url-loader": "^0.5.8",
-        "vue-loader": "^11.3.4",
-        "vue-style-loader": "^2.0.5",
-        "vue-template-compiler": "^2.3.4",
-        "webpack": "^2.7.0",
-        "webpack-bundle-analyzer": "^2.2.1",
-        "webpack-dev-middleware": "^1.10.0",
-        "webpack-hot-middleware": "^2.18.0",
-        "webpack-merge": "^4.1.0"
-    },
-    "engines": {
-        "node": ">= 8.0.0",
-        "npm": ">= 4.0.0"
+    "parserOptions": {
+      "parser": "babel-eslint"
     }
+  },
+  "postcss": {
+    "plugins": {
+      "autoprefixer": {}
+    }
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not ie <= 8"
+  ]
 }
