@@ -1,70 +1,9 @@
-import VueHtml5Editor from 'vue-html5-editor'
-const VueEditor = new VueHtml5Editor({
-    name: "vue-html5-editor",
-    icons: {
-        text: "fa fa-pencil",
-        color: "fa fa-paint-brush",
-        font: "fa fa-font",
-        align: "fa fa-align-justify",
-        list: "fa fa-list",
-        link: "fa fa-chain",
-        unlink: "fa fa-chain-broken",
-        tabulation: "fa fa-table",
-        image: "fa fa-file-image-o",
-        hr: "fa fa-minus",
-        eraser: "fa fa-eraser",
-        undo: "fa-undo fa",
-        "full-screen": "fa fa-arrows-alt"
-    },
-    image: {
-        sizeLimit: 512 * 1024,
-        // 上传参数,默认把图片转为base64而不上传
-        // upload config,default null and convert image to base64
-        upload: {
-            url: null,
-            headers: {},
-            params: {},
-            fieldName: {}
-        },
-        // 压缩参数,默认使用localResizeIMG进行压缩,设置为null禁止压缩
-        // compression config,default resize image by localResizeIMG (https://github.com/think2011/localResizeIMG)
-        // set null to disable compression
-        compress: {
-            width: 1000,
-            height: 1000,
-            quality: 80
-        },
-        // 响应数据处理,最终返回图片链接
-        // handle response data，return image url
-        uploadHandler(responseText){
-            //default accept json data like  {ok:false,msg:"unexpected"} or {ok:true,data:"image url"}
-            let json = JSON.parse(responseText);
-            if (!json.ok) {
-                alert(json.msg)
-            } else {
-                return json.data
-            }
-        }
-    },
-    hiddenModules: ["info"],
-    language: "zh-cn",
-    visibleModules: [
-        "text",
-        "color",
-        "font",
-        "align",
-        "list",
-        "link",
-        "unlink",
-        "tabulation",
-        "image",
-        "hr",
-        "eraser",
-        "undo",
-        "full-screen"
-    ]
-});
-module.exports = {
-    upFile: require('./upFile.vue'),
-    VueEditor
-};
+import { quillEditor } from 'vue-quill-editor'
+import 'quill/dist/quill.core.css'
+import 'quill/dist/quill.snow.css'
+import 'quill/dist/quill.bubble.css'
+export default {
+  quillEditor,
+  upFile: require('./upFile.vue').default,
+  Sidebar: require('./Sidebar.vue').default
+}
