@@ -1,10 +1,10 @@
-require("babel-core/register")({
-  "presets": [
-    ["env", {
-      "targets": {
-        "node": true
+require('babel-core/register')({
+  'presets': [
+    ['env', {
+      'targets': {
+        'node': true
       }
     }]
   ]
-});
-require("./app.js");
+})
+require('./app.js')
