@@ -5,8 +5,7 @@
 </template>
 
 <script>
-    export default {
-    };
+export default {}
 </script>
 
 <style lang="less" scoped>
